Minor cleanup and added stub for Protected change level
